renamed hooke.ts to withTypes.ts and moved it to 'store' dir
