+++ src/components/FilmItem.tsx
@@ -1,6 +1,6 @@
 import FilmService from "../api/FilmService";
 import { selectUser } from "../features/user/userSlice";
-import { useAppSelector } from "../hooks/hooks";
+import { useAppSelector } from "../store/withTypes";
 import { useFetching } from "../hooks/useFetching";
 import { Film } from "./FilmsDashboard";
